Make OptionTabs component
