+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/EditCodeList/EditCodeListReference.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import type { IGenericEditComponent } from '../../../componentConfig';
+import type { IGenericEditComponent } from '../../../../componentConfig';
 import { useTranslation, Trans } from 'react-i18next';
 
 import { altinnDocsUrl } from 'app-shared/ext-urls';
 import { StudioParagraph, StudioTextfield } from '@studio/components';
-import type { SelectionComponentType } from '../../../../../types/FormComponent';
+import type { SelectionComponentType } from '../../../../../../types/FormComponent';
 
 export function EditCodeListReference<T extends SelectionComponentType>({
   component,
