+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/EditCodeList/EditCodelistReference.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { screen } from '@testing-library/react';
 import { EditCodeListReference } from './EditCodeListReference';
-import { renderWithProviders } from '../../../../../testing/mocks';
+import { renderWithProviders } from '../../../../../../testing/mocks';
 import { ComponentType } from 'app-shared/types/ComponentType';
-import type { FormComponent } from '../../../../../types/FormComponent';
+import type { FormComponent } from '../../../../../../types/FormComponent';
 import { textMock } from '@studio/testing/mocks/i18nMock';
 import userEvent from '@testing-library/user-event';
