+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/EditManualOptions/EditManualOptions.test.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { screen } from '@testing-library/react';
 import { EditManualOptions } from './EditManualOptions';
-import { renderWithProviders } from '../../../../../testing/mocks';
+import { renderWithProviders } from '../../../../../../testing/mocks';
 import { textMock } from '@studio/testing/mocks/i18nMock';
 import { ComponentType } from 'app-shared/types/ComponentType';
-import type { FormItem } from '../../../../../types/FormItem';
+import type { FormItem } from '../../../../../../types/FormItem';
 import type { ServicesContextProps } from 'app-shared/contexts/ServicesContext';
-import type { FormComponent } from '../../../../../types/FormComponent';
+import type { FormComponent } from '../../../../../../types/FormComponent';
 import userEvent from '@testing-library/user-event';
 
 const mockComponent: FormComponent<ComponentType.RadioButtons> = {
