+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/EditManualOptions/EditOption/EditOption.module.css
