+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/EditManualOptions/EditOption/OptionValue/OptionValue.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { Option } from 'app-shared/types/Option';
-import { useTextResourcesSelector } from '../../../../../hooks';
-import { textResourceByLanguageAndIdSelector } from '../../../../../selectors/textResourceSelectors';
+import { useTextResourcesSelector } from '../../../../../../../../hooks';
+import { textResourceByLanguageAndIdSelector } from '../../../../../../../../selectors/textResourceSelectors';
 import { DEFAULT_LANGUAGE } from 'app-shared/constants';
 import { StudioCodeFragment } from '@studio/components';
 import type { ITextResource } from 'app-shared/types/global';
