+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/hooks/useCodeListButtonValue.ts
