+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/hooks/useCodeListEditorTexts.ts
