+++ frontend/packages/ux-editor/src/components/config/editModal/EditOptions/OptionTabs/index.ts
@@ -1,0 +1,1 @@
+export { OptionTabs } from './OptionTabs';
